move ParseUtils to util package
