+++ src/main/java/org/opensearch/ad/model/ADTask.java
@@ -26,7 +26,7 @@ import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.timeseries.annotation.Generated;
 import org.opensearch.timeseries.model.DateRange;
-import org.opensearch.timeseries.model.ParseUtils;
+import org.opensearch.timeseries.util.ParseUtils;
 
 import com.google.common.base.Objects;
