+++ src/main/java/org/opensearch/ad/transport/DeleteAnomalyResultsTransportAction.java
@@ -14,8 +14,8 @@ package org.opensearch.ad.transport;
 import static org.opensearch.ad.constant.ADCommonMessages.FAIL_TO_DELETE_AD_RESULT;
 import static org.opensearch.ad.settings.AnomalyDetectorSettings.FILTER_BY_BACKEND_ROLES;
 import static org.opensearch.ad.util.RestHandlerUtils.wrapRestActionListener;
-import static org.opensearch.timeseries.model.ParseUtils.addUserBackendRolesFilter;
-import static org.opensearch.timeseries.model.ParseUtils.getUserContext;
+import static org.opensearch.timeseries.util.ParseUtils.addUserBackendRolesFilter;
+import static org.opensearch.timeseries.util.ParseUtils.getUserContext;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
