+++ src/main/java/org/opensearch/ad/transport/handler/ADSearchHandler.java
@@ -14,9 +14,9 @@ package org.opensearch.ad.transport.handler;
 import static org.opensearch.ad.constant.ADCommonMessages.FAIL_TO_SEARCH;
 import static org.opensearch.ad.settings.AnomalyDetectorSettings.FILTER_BY_BACKEND_ROLES;
 import static org.opensearch.ad.util.RestHandlerUtils.wrapRestActionListener;
-import static org.opensearch.timeseries.model.ParseUtils.addUserBackendRolesFilter;
-import static org.opensearch.timeseries.model.ParseUtils.getUserContext;
-import static org.opensearch.timeseries.model.ParseUtils.isAdmin;
+import static org.opensearch.timeseries.util.ParseUtils.addUserBackendRolesFilter;
+import static org.opensearch.timeseries.util.ParseUtils.getUserContext;
+import static org.opensearch.timeseries.util.ParseUtils.isAdmin;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
