+++ src/main/java/org/opensearch/timeseries/model/DateRange.java
@@ -24,6 +24,7 @@ import org.opensearch.core.xcontent.ToXContentObject;
 import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.timeseries.annotation.Generated;
+import org.opensearch.timeseries.util.ParseUtils;
 
 import com.google.common.base.Objects;
